Draw seems fine now
After all the problem was not on my code. Keras has an algorithimic choice that
I was not awere of. It indiscriminately collapse all the dimmensions of the
output (but the last one) with the first (samples) dimension. Since I was using
images (samples, channels, rows, cols) as desired, it messed up my gradients.